backend/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
python-iso639==2024.4.27
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==0.6.23
SQLAlchemy==2.0.31
tqdm==4.66.4
llama-index==0.10.58
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.58
llama-index-embeddings-ollama==0.1.2
llama-index-embeddings-openai==0.1.11
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.8
llama-index-vector-stores-qdrant==0.2.14
Markdown==3.6
qdrant-client==1.10.1
fastapi==0.110.3
celery==5.4.0
moviepy==1.0.3
pydub==0.25.1
google-generativeai==0.7.2
python-dotenv==1.0.1
chromadb==0.5.5
redis==5.0.8
easydict==1.13
